Shifted weights into the column, because weighting applies to the column, wrote tests and created a new stats class to handle more complicated stats
