+++ CHAID/node.py
@@ -27,7 +27,7 @@ class Node(object):
     is_terminal : boolean
         Whether the node is terminal
     """
-    def __init__(self, choices=None, split=None, indices=None, node_id=0, parent=None, dep_v=None, is_terminal=False, weights=None):
+    def __init__(self, choices=None, split=None, indices=None, node_id=0, parent=None, dep_v=None, is_terminal=False):
         indices = [] if indices is None else indices
         self.choices = list(choices or [])
         self.split = split or Split(None, None, None, None, 0)
@@ -37,7 +37,6 @@ class Node(object):
         self.dep_v = dep_v
         self._members = None
         self.is_terminal = is_terminal
-        self.weights = weights
 
     def __hash__(self):
         return hash(self.__dict__)
@@ -76,11 +75,11 @@ class Node(object):
             for member in metadata.values():
                 self._members[member] = 0
 
-            if self.weights is None:
+            if dep_v.weights is None:
                 counts = np.transpose(np.unique(dep_v.arr, return_counts=True))
             else:
                 counts = np.array([
-                    [i, self.weights[dep_v.arr == i].sum()] for i in set(dep_v.arr)
+                    [i, dep_v.weights[dep_v.arr == i].sum()] for i in set(dep_v.arr)
                 ])
 
             self._members.update((metadata[k], v) for k, v in counts)
